refactor(githubSync): split 1602-line module into pull/push/classify (#77)
`src/utils/githubSync.ts` had grown to ~1602 lines and folded the pull
orchestration, push orchestration, the PullClassification kinds, the
zipball fast path, and every shared helper into one file. Split it
into focused submodules under `src/utils/githubSync/`:

- `internal.ts`     — path computation (notePath/pushPath/
  collectLocalFolderRepoPaths), MIME guessing, note serialization
  + parser (serializeNote/normalizeForPush/parseNote/parseInlineArrayField),
  the canonical-form equality helper, and the maybeEncryptForPush /
  maybeDecryptFromPull vault-key wrappers. Shared by both halves.
- `syncClassify.ts` — the `PullClassification` discriminated union and
  the `PullOutcome` shape. Was a ~250-line type block embedded in the
  pull file; it now lives on its own so the kinds (the six core +
  attachment/folder/settings extensions) can be referenced and
  extended without touching the orchestration files.
- `syncPull.ts`     — `pullFromGitHub`, `pullFromZipball`, and the
  zipball attachment side-channel cache (`takeZipballAttachmentBytes`).
- `syncPush.ts`     — `syncToGitHub`, the in-memory uploaded-blob-SHA
  cache, and the `_resetUploadedShaCache` test hook.

The top-level `src/utils/githubSync.ts` is now a thin barrel that
re-exports the public surface — every external import path
(`@/utils/githubSync` / `../utils/githubSync` / `./githubSync`) keeps
working untouched. No test or downstream import was updated.

Behavior-preserving: 2378/2378 jest tests pass, `npm run lint` clean,
`npx tsc --noEmit` clean.
